requirements.txt
annotated-types==0.6.0
beautifulsoup4==4.12.2
bert-extractive-summarizer==0.10.1
blis==0.7.11
bs4==0.0.1
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
cymem==2.0.8
feedparser==6.0.10
filelock==3.12.4
fsspec==2023.9.2
gensim==4.3.2
huggingface-hub==0.17.3
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
MarkupSafe==2.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.1
numpy==1.26.0
packaging==23.2
pathy==0.10.2
preshed==3.0.9
pydantic==2.4.2
pydantic_core==2.10.1
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
safetensors==0.4.0
scikit-learn==1.3.1
scipy==1.11.3
sgmllib3k==1.0.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.2.1
threadpoolctl==3.2.0
tokenizers==0.14.1
torch==2.1.0
tqdm==4.66.1
transformers==4.34.0
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.6
wasabi==1.1.2
weasel==0.3.2
